qla2xxx: Stop trying to load qla82xx firmware via request_firmware if ql2xfwloadbin=2 is set.
During a chip reset on a qla82xx CNA, we will try to load the firmware via the
request_firmware interface when the chip is coming back up and the
ql2xfwloadbin=2 driver parameter is set.  Remove the logic that does this
so the firmware is simply loaded from flash.