qla2xxx: Display FCP_CMND priority on update.
Some of the targets don't like priority value set in FCP_CMND and that leads to
invalid FCP_CMND and no luns are discovered. It's hard to find that the problem is
that the priority is set in FCP_CMND so add debug statements when we set the
priority in a FCP_CMND struct so that we can see if this is happening.

JIRA Key: V2632FC-33