qla2xxx: Add dynamic logging.
This patch adds the dynamic logging framework to the qla2xxx driver.
The user will be able to change the logging levels on the fly i.e.
without load/unload of the driver. This also enables logging to be enabled for
a particular section of the driver such as initialization, device discovery etc.

JIRA Key: V2632FC-64