qla2xxx: Don't process RSCNs for a vport on the same physical adapter.
Currently, the driver is processes RSCNs for each new NPIV ports that is created.
Processing the RSCN includes a name server query to see what has changed at the
name server side.  The name server query is performed by the physical port and
each virtual port on the physical adapter (since the RSCN is passed to each
virtual port for processing).  As the number of virtual ports being created
increases, this causes a lot of traffic and busies the firmware.

Processing the RSCN for a virtual port we already have a priori knowledge of is
not necessary so check the 24-bit fabric ID of the RSCN entry and skip
processing it if the RSCN is for a virtual port we already know about.

JIRA Key: V2632FC-322
ER: 101023

Acked-by: Saurav Kashyap <saurav.kashyap@qlogic.com>
Acked-by: Atul Deshmukh <atul.deskmukh@qlogic.com>