qla2xxx: Fix request queue null dereference.
If an invalid IOCB is returned on the response queue then the index into the
request queue map could be invalid and could return to us a bogus value. This
could cause us to try to deference an invalid pointer and cause an exception.

If we encounter this condition, simply return as no context can be established
for this response.

JIRA Key: V2632FC-465

Acked-by: Joe Carnuccio <joe.carnuccio@qlogic.com>
Acked-by: Atul Deshmukh <atul.deshmukh@qlogic.com>