qla2xxx: Disable adapter when we encounter a PCI disconnect.
If we become disconnected from the PCI bus/PCIe fabric, there can be long delays
in register reads which can cause erroneous decisions to be made and cause a
soft lockup if a lock is held too long. As a preventative measure, check for a
disconnection (register reads that return -1) and then disable the board if we
find ourselves in this condition. For now, check in our interrupt handlers and
the per adapter one second timer.

JIRA Key: V2632FC-364

Acked-by: Giridhar Malavali <giridhar.malavali@qlogic.com>
Acked-by: Saurav Kashyap <saurav.kashyap@qlogic.com>